requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
httpx>=0.24.1
google-genai>=1.0.0
